ref: move Calendar props into CalendarBase directory
so range calendar can reuse shared props
